Finished moderator analysis and regular analysis (need some textual work)
Moderator and regular analysis are split up in two folders.
Cache and files for Rmd document are not pushed.

The analyses contain plots connecting the values for beta and tau.
Percentiles for beta and tau separately are checked.

Need some work on text of regular analysis as most comes from moderator analysis section.