Mitigated true value parameters from true values file to R package.
The true values were previously defined in both the simulation script,
the true values file (MAvsIBMA_Act_true_values.R) and the files for analysis.
To reduce amount of copying, we are moving the definition of global true values
to the R package fMRIGI.
This commit updates the function trueMCvalues with the values defined in the
true values file.