Adding reports with FLAME and FLMER for second and third level variances.
So we generate time series for just one voxel, and have a full mixed model.
That is: random slopes and intercepts for subjects (second level) and studies
(third levels).

We use FLAME for the group level models. Finally we use both R (LMER) and FSL
to estimate the parameters of the GLM at the study level.
We find good coverages and variance of the estimates seems ok.
However the estimate of the variance for between-study variability is off.

Will now go back to original scripts and see what is going wrong there!