Made compatible with signal in one voxel and new variable names. Checked code. Removed one bug.
There was one bug in calculating the coverage for GLM where I used the upper CI bound of MA instead of GLM.
Now removed code that was used when one signal was given in one voxel and we used smoothing
of the true values.
Instead we have a raw signal in each voxel now.
Checked code. Then made variable names compatible with new names.